Add macOS to CI (#17)
* Add DEBUG option to fallback to using allocator, so that valgrind can catch memory bugs

* Adjust dup/drop for fib.c

* Fix ssm_sv_init macro

Previously, this "uninitialized" the mm header field in the SV, wreaking
all sorts of havoc

* Perform valgrind testing without memory pools

* Add macos platform, remove redundant dependency installation

* Add valgrind list test

* Fix syntax error in workflow file

* Add support for different compilers

* Test with different compilers

* Add note about symlink to clang from gcc on macos

* Add cast to suppress warning

+++ examples/ssm-examples.h
@@ -128,7 +128,8 @@ int main(int argc, char *argv[]) {
   while (ssm_next_event_time() != SSM_NEVER && ssm_now() < stop_at)
     ssm_tick();
 
-  printf("%s: simulated %lu seconds\n", prog, ssm_now() / SSM_SECOND);
+  printf("%s: simulated %lu seconds\n", prog,
+         (unsigned long)ssm_now() / SSM_SECOND);
 
   ssm_program_exit();
